Fold ResponseBuilder into generic Responder class, pushing actual writing responsibility to the ClientWorker. More refactoring.
