+++ src/main/java/javaserver/ClientWorker.java
@@ -8,18 +8,17 @@ public class ClientWorker implements Runnable {
 	public SocketWriter writer;
 	private Responder responder;
 
-	public ClientWorker(ServerFactory serverFactory, Reader reader, Responder responder, SocketWriter writer) {
+	public ClientWorker(ServerFactory serverFactory, Reader reader, SocketWriter writer) {
 		this.serverFactory = serverFactory;
 		this.reader = reader;
-		this.responder = responder;
 		this.writer = writer;
 	}
 
 	public void run() {
 		String rawRequest = getRequest();
 		Request request = new Request(RequestParser.getRequestMethod(rawRequest), RequestParser.getRequestURI(rawRequest));
+		this.responder = new Responder(request);
 		respond(request);
-		writer.closeOutputStream();
 	}
 
 	private String getRequest() {
@@ -34,9 +33,10 @@ public class ClientWorker implements Runnable {
 
 	private void respond(Request request) {
 		try {
-			responder.respond(request, serverFactory, writer);
+			Response response = responder.getResponse();
+			writer.respond(response.formatResponse());
 		} catch (IOException e) {
-			System.out.println("ClientWorker was unable to respond to client socket");
+			System.out.println("ClientWorker unable to get response from Responder or write Response with SocketWriter");
 		}
 	}
 }
