+++ src/main/java/javaserver/Responder.java
@@ -1,21 +1,68 @@
 package javaserver;
 
-import java.io.IOException;
+import java.io.UnsupportedEncodingException;
 
 public class Responder {
-	Request request;
-	ServerFactory serverFactory;
-
-	public void respond(Request request, ServerFactory serverFactory, SocketWriter writer) throws IOException {
-		this.serverFactory = serverFactory;
+	private Request request;
+	protected Response response;
+	
+	Responder(Request request) {
 		this.request = request;
-		Response response = getResponse();
-		String formattedResponse = response.getFormattedResponse();
-		writer.respond(formattedResponse);
+	}
+	
+	public Response getResponse() {
+		this.response = new Response();
+		setResponseData();
+		return this.response;
 	}
 
-	private Response getResponse() {
-		ResponseBuilder responseBuilder = serverFactory.createResponseBuilder(request);
-		return responseBuilder.createResponse();
+	private void setResponseData() {
+		response.setResponseCode(getResponseCode());
+		response.setHeader(getResponseHeader());
+		response.setBody(getResponseBody());
+	}
+
+	private String getResponseCode() {
+		if (request.isOK()) {
+			return HTTPStatusCodes.TWO_HUNDRED;
+		} else {
+			return HTTPStatusCodes.FOUR_OH_FOUR;
+		}
+	}
+
+	private String getResponseHeader() {
+		String header = new String();
+		if (request.getURI().equals("/method_options")) {
+			header += "Allow: ";
+			header += String.join(",", request.routeOptions());
+		}
+		return header;
+	}
+
+	private String getResponseBody() {
+		String body = new String();
+		if (request.hasVariableParams()) {
+			String[] allParams = separateParameters();
+			for (int i = 0; i < allParams.length; i++) {
+				body += decode(allParams[i]) + System.lineSeparator();
+			}
+		}
+		return body;
+	}
+
+	private String[] separateParameters() {
+		String params = request.getURI().split("/parameters?.")[1];
+		params = params.replace("=", " = ");
+		return params.split("&");
+	}
+	
+	private String decode(String parameter) {
+		try {
+			String encoding = "UTF-8";
+			parameter = java.net.URLDecoder.decode(parameter, encoding);
+		} catch (UnsupportedEncodingException e) {
+			System.out.println("ResponseBuilder could not decode one or more of the request's parameters");
+		}
+		return parameter;
 	}
 }
