+++ src/main/java/javaserver/SocketWriter.java
@@ -15,6 +15,7 @@ public class SocketWriter {
 	public void respond(String response) throws IOException {
 		writingMechanism.writeBytes(response);
 		this.latestResponse = response;
+		closeOutputStream();
 	}
 
 	public void closeOutputStream() {
